requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-environ==0.11.2
djangorestframework==3.14.0
dr-scaffold==2.1.2
idna==3.4
inflect==7.0.0
isort==5.12.0
mysqlclient==2.2.0
pydantic==2.4.2
pydantic_core==2.10.1
PyMySQL==1.1.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.5
